feat: switch to lazy
