+++ lua/arnar/plugins/chatgpt.lua
@@ -1,0 +1,13 @@
+return {
+  "jackMort/ChatGPT.nvim",
+  config = function()
+    require("chatgpt").setup({
+      api_key_cmd = "lpass show --password 'chatgpt api key'",
+    })
+  end,
+  dependencies = {
+    "MunifTanjim/nui.nvim",
+    "nvim-lua/plenary.nvim",
+    "nvim-telescope/telescope.nvim"
+  }
+}
