+++ lua/arnar/plugins/fugitive.lua
@@ -1,0 +1,6 @@
+return {
+  "tpope/vim-fugitive",
+  config = function()
+    vim.keymap.set('n', '<c-g>', ':Git', { noremap = true, silent = false })
+  end,
+}
