+++ lua/arnar/plugins/init.lua
@@ -1,0 +1,30 @@
+return {
+  "folke/which-key.nvim",
+  { "folke/neoconf.nvim",                       cmd = "Neoconf" },
+  "folke/neodev.nvim",
+  "nvim-lua/popup.nvim",       -- An implementation of the Popup API from vim in Neovim
+  "nvim-lua/plenary.nvim",     -- Useful lua functions used ny lots of plugins
+
+  "gruvbox-community/gruvbox", -- colorsheme
+  "Shatur/neovim-ayu",
+  'folke/tokyonight.nvim',
+
+  -- Should be a part of vim
+  "tpope/vim-surround",
+  "tpope/vim-abolish",
+  "tpope/vim-commentary",
+  "vim-scripts/ReplaceWithRegister",
+  "bronson/vim-visual-star-search",
+
+  "christoomey/vim-tmux-navigator", -- navigate between vim and tmux panes
+  'nvim-treesitter/nvim-treesitter-context',
+  'williamboman/mason.nvim',
+  'neovim/nvim-lspconfig',
+  'hrsh7th/cmp-nvim-lsp',
+  'hrsh7th/cmp-buffer',
+  'hrsh7th/cmp-path',
+  'hrsh7th/cmp-cmdline',
+  { "L3MON4D3/LuaSnip",                         version = "v1.*" },
+  { 'nvim-telescope/telescope-fzf-native.nvim', build = 'make' },
+  { 'norcalli/nvim-colorizer.lua' },
+}
