+++ lua/arnar/plugins/treesitter.lua
@@ -1,0 +1,16 @@
+return {
+  "nvim-treesitter/nvim-treesitter",
+  build = ":TSUpdate",
+  config = function()
+    local configs = require 'nvim-treesitter.configs'
+    configs.setup {
+      ensure_installed = { "lua", "vim", "javascript", "typescript", "tsx" },
+      highlight = { -- enable highlighting
+        enable = true,
+      },
+      indent = {
+        enable = false, -- default is disabled anyways
+      }
+    }
+  end,
+}
