--- lua/plugins/chatgpt.lua
@@ -1,3 +1,0 @@
-require("chatgpt").setup({
-    api_key_cmd = "lpass show --password 'chatgpt api key'",
-})
