--- lua/plugins/indent_blankline.lua
@@ -1,1 +1,0 @@
-require("ibl").setup { }
