--- lua/plugins/prettier.lua
@@ -1,7 +1,0 @@
-vim.g.neoformat_try_node_exe = 1
-
-vim.api.nvim_create_autocmd('BufWritePre', {
-  pattern = {'*.js', '*.ts', '*.tsx'},
-  command = 'Neoformat prettier'
-})
-
